requirements_dev.txt
-r requirements_base.txt
-e .
loremipsum==1.0.5
pytest==2.6.4
pytest-django==2.7.0
